arm64: fix endianness annotation for 'struct jit_ctx' and friends
struct jit_ctx::image is used the store a pointer to the jitted
intructions, which are always little-endian. These instructions
are thus correctly converted from native order to little-endian
before being stored but the pointer 'image' is declared as for
native order values.

Fix this by declaring the field as __le32* instead of u32*.
Same for the pointer used in jit_fill_hole() to initialize
the image.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Change-Id: I8f00a950473046976040fed6f70c4d6104051fc3
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>